Removes images.js and images task
