+++ uncompressed/apple-touch-icon.png
