+++ uncompressed/favicon-32x32.png
